% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibility_prefs.R
\name{prefs_improve_reproducibility}
\alias{prefs_improve_reproducibility}
\title{Improve reproducibility of RStudio}
\usage{
prefs_improve_reproducibility()
}
\description{
prefs_improve_reproducibility() is designed to improve the reproducibility of
your RStudio environment. It changes the following preferences:
\itemize{
\item "save_workspace": "never" - this prevents RStudio asking if you want to
save
\item "always_save_history": false - this prevent RStudio saving a .RHistory file
\item "load_workspace": false - thi prevent RStudio loading .RData files
automatically on project loads
}
}
