% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prefs.R
\name{prefs_reset_to_defaults}
\alias{prefs_reset_to_defaults}
\title{Restore default preferences}
\usage{
prefs_reset_to_defaults(improve_code_reproducibility = TRUE)
}
\description{
\code{prefs_reset_to_defaults()} resets RStudio preferences to defaults. You are
\strong{strongly} recommend to run \code{prefs_improve_reproducibility()} after running
prefs_reset_to_defaults().
}
